+++ msamp/optim/__init__.py
@@ -5,6 +5,7 @@
 
 from msamp.optim.optimizer import LBOptimizer
 from msamp.optim.adamw_base import LBAdamWBase
-from msamp.optim.adamw import LBAdamW, LBAdam, DsLBAdam
+from msamp.optim.adamw import LBAdamW
+from msamp.optim.adam import LBAdam, DSAdam
 
-__all__ = ['LBOptimizer', 'LBAdamWBase', 'LBAdamW', 'LBAdam', 'DsLBAdam']
+__all__ = ['LBOptimizer', 'LBAdamWBase', 'LBAdamW', 'LBAdam', 'DSAdam']
